# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type TransactionType = typeof TransactionType[keyof typeof TransactionType];


export const TransactionType = {
  income: 'income',
  expense: 'expense',
} as const;

export interface Transaction {
  id: number;
  amount: number;
  category: string;
  description: string;
  date: string;
  type: TransactionType;
  /** @nullable */
  aiConfidence: number | null;
}

export type TransactionInputType = typeof TransactionInputType[keyof typeof TransactionInputType];


export const TransactionInputType = {
  income: 'income',
  expense: 'expense',
} as const;

export interface TransactionInput {
  /** @minimum 0.01 */
  amount: number;
  /** @minLength 1 */
  category: string;
  /** @minLength 1 */
  description: string;
  date: string;
  type: TransactionInputType;
  /** @nullable */
  aiConfidence?: number | null;
}

export type TransactionUpdateType = typeof TransactionUpdateType[keyof typeof TransactionUpdateType];


export const TransactionUpdateType = {
  income: 'income',
  expense: 'expense',
} as const;

export interface TransactionUpdate {
  /** @minimum 0.01 */
  amount?: number;
  /** @minLength 1 */
  category?: string;
  /** @minLength 1 */
  description?: string;
  date?: string;
  type?: TransactionUpdateType;
  /** @nullable */
  aiConfidence?: number | null;
}

export interface Budget {
  id: number;
  category: string;
  limitAmount: number;
  spentAmount: number;
  month: string;
  color: string;
}

export interface BudgetInput {
  /** @minimum 0.01 */
  limitAmount: number;
  /**
     * @minLength 7
     * @maxLength 7
     */
  month: string;
}

export interface CategorySpend {
  category: string;
  amount: number;
  color: string;
}

export interface WeeklySpend {
  day: string;
  amount: number;
}

export interface DashboardSummary {
  totalBalance: number;
  monthlySpend: number;
  monthlyIncome: number;
  savingsRate: number;
  spendChange: number;
  spendByCategory: CategorySpend[];
  weeklySpend: WeeklySpend[];
  recentTransactions: Transaction[];
}

export type InsightImpact = typeof InsightImpact[keyof typeof InsightImpact];


export const InsightImpact = {
  positive: 'positive',
  neutral: 'neutral',
  warning: 'warning',
} as const;

export interface Insight {
  id: string;
  title: string;
  body: string;
  category: string;
  impact: InsightImpact;
  createdAt: string;
}

export interface CategorizeInput {
  /** @minLength 1 */
  description: string;
}

export interface CategorizeResult {
  category: string;
  /**
     * @minimum 0
     * @maximum 1
     */
  confidence: number;
}

export type ListTransactionsParams = {
search?: string;
category?: string;
type?: ListTransactionsType;
sort?: ListTransactionsSort;
};

export type ListTransactionsType = typeof ListTransactionsType[keyof typeof ListTransactionsType];


export const ListTransactionsType = {
  income: 'income',
  expense: 'expense',
} as const;

export type ListTransactionsSort = typeof ListTransactionsSort[keyof typeof ListTransactionsSort];


export const ListTransactionsSort = {
  date_desc: 'date_desc',
  date_asc: 'date_asc',
  amount_desc: 'amount_desc',
  amount_asc: 'amount_asc',
} as const;