# lib/api-zod/src/generated/types/budget.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Budget {
  id: number;
  category: string;
  limitAmount: number;
  spentAmount: number;
  month: string;
  color: string;
}