# lib/api-zod/src/generated/types/budgetInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface BudgetInput {
  /** @minimum 0.01 */
  limitAmount: number;
  /**
     * @minLength 7
     * @maxLength 7
     */
  month: string;
}