# lib/api-zod/src/generated/types/categorizeResult.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CategorizeResult {
  category: string;
  /**
     * @minimum 0
     * @maximum 1
     */
  confidence: number;
}