# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CategorySpend } from './categorySpend';
import type { Transaction } from './transaction';
import type { WeeklySpend } from './weeklySpend';

export interface DashboardSummary {
  totalBalance: number;
  monthlySpend: number;
  monthlyIncome: number;
  savingsRate: number;
  spendChange: number;
  spendByCategory: CategorySpend[];
  weeklySpend: WeeklySpend[];
  recentTransactions: Transaction[];
}