# lib/api-zod/src/generated/types/transactionUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TransactionUpdateType } from './transactionUpdateType';

export interface TransactionUpdate {
  /** @minimum 0.01 */
  amount?: number;
  /** @minLength 1 */
  category?: string;
  /** @minLength 1 */
  description?: string;
  date?: Date;
  type?: TransactionUpdateType;
  /** @nullable */
  aiConfidence?: number | null;
}